version 1

+++ Python/PythonOperator/CountingNotes.py
@@ -1,4 +1,4 @@
-amount_input = float(input("Enter an amount within the multiplication of 100, 50 or 10 taka to withdraw: "))
+amount_input = int(input("Enter an amount within the multiplication of 100, 50 or 10 taka to withdraw: "))
 print(amount_input)
 
 note1 = amount_input // 100
